requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiofiles=22.1.0=py311h06a4308_0
aiosqlite=0.18.0=py311h06a4308_0
anyio=3.5.0=py311h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py311h5eee18b_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=23.1.0=py311h06a4308_0
babel=2.11.0=py311h06a4308_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=py311h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py311h5eee18b_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.7.22=py311h06a4308_0
cffi=1.15.1=py311h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.2=py311h06a4308_0
contourpy=1.0.5=py311hdb19cb5_0
cryptography=41.0.3=py311hdda0065_0
cycler=0.11.0=pyhd3eb1b0_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py311h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py311h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.5.0=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py311h06a4308_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gnutls=3.6.15=he1e5248_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=58.2=he6710b0_3
idna=3.4=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipdb=0.13.13=pypi_0
ipykernel=6.25.0=py311h92b7b1e_0
ipython=8.15.0=py311h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.0.4=py311h06a4308_0
jedi=0.18.1=py311h06a4308_1
jinja2=3.1.2=py311h06a4308_0
joblib=1.2.0=py311h06a4308_0
jpeg=9e=h5eee18b_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.17.3=py311h06a4308_0
jupyter=1.0.0=py311h06a4308_8
jupyter_client=7.4.9=py311h06a4308_0
jupyter_console=6.6.3=py311h06a4308_0
jupyter_core=5.3.0=py311h06a4308_0
jupyter_events=0.6.3=py311h06a4308_0
jupyter_server=1.23.4=py311h06a4308_0
jupyter_server_fileid=0.9.0=py311h06a4308_0
jupyter_server_ydoc=0.8.0=py311h06a4308_1
jupyter_ydoc=0.2.4=py311h06a4308_0
jupyterlab=3.6.3=py311h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.22.0=py311h06a4308_0
jupyterlab_widgets=3.0.5=py311h06a4308_0
kiwisolver=1.4.4=py311h6a678d5_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=default_hc6dbbc7_1
libclang13=14.0.6=default_he11475f_1
libcups=2.4.2=h2d74bed_1
libdeflate=1.17=h5eee18b_1
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libllvm14=14.0.6=hdb19cb5_3
libpng=1.6.39=h5eee18b_0
libpq=12.15=hdbd6064_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hcbfbd50_0
libxslt=1.1.37=h2085143_0
lxml=4.9.3=py311hdbbb534_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py311h5eee18b_0
matplotlib=3.7.2=py311h06a4308_0
matplotlib-base=3.7.2=py311ha02d727_0
matplotlib-inline=0.1.6=py311h06a4308_0
mistune=0.8.4=py311h5eee18b_1000
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
munkres=1.1.4=py_0
mysql=5.7.24=h721c034_2
nbclassic=0.5.5=py311h06a4308_0
nbclient=0.5.13=py311h06a4308_0
nbconvert=6.5.4=py311h06a4308_0
nbformat=5.9.2=py311h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py311h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py311h06a4308_0
notebook=6.5.4=py311h06a4308_1
notebook-shim=0.2.2=py311h06a4308_0
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numpy=1.26.0=py311h08b1b3b_0
numpy-base=1.26.0=py311hf175353_0
opencv-python=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.12=h7f8727e_0
packaging=23.1=py311h06a4308_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py311h6a678d5_1
pip=23.2.1=py311h06a4308_0
platformdirs=3.10.0=py311h06a4308_0
ply=3.11=py311h06a4308_0
prometheus_client=0.14.1=py311h06a4308_0
prompt-toolkit=3.0.36=py311h06a4308_0
prompt_toolkit=3.0.36=hd3eb1b0_0
psutil=5.9.0=py311h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py311h06a4308_1
pyopenssl=23.2.0=py311h06a4308_0
pyparsing=3.0.9=py311h06a4308_0
pyqt=5.15.7=py311h6a678d5_0
pyqt5-sip=12.11.0=py311h6a678d5_0
pyrsistent=0.18.0=py311h5eee18b_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.5=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py311h06a4308_0
python-json-logger=2.0.7=py311h06a4308_0
pytorch=2.0.1=py3.11_cpu_0
pytorch-mutex=1.0=cpu
pytz=2023.3.post1=py311h06a4308_0
pyyaml=6.0=py311h5eee18b_1
pyzmq=23.2.0=py311h6a678d5_0
qt-main=5.15.2=h7358343_9
qt-webengine=5.15.9=h9ab4d14_7
qtconsole=5.4.2=py311h06a4308_0
qtpy=2.2.0=py311h06a4308_0
qtwebkit=5.212=h3fafdc1_5
readline=8.2=h5eee18b_0
requests=2.31.0=py311h06a4308_0
rfc3339-validator=0.1.4=py311h06a4308_0
rfc3986-validator=0.1.1=py311h06a4308_0
scikit-learn=1.2.2=py311h6a678d5_1
scipy=1.11.1=py311h08b1b3b_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=68.0.0=py311h06a4308_0
sip=6.6.2=py311h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py311h06a4308_1
soupsieve=2.4=py311h06a4308_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.11.1=py311h06a4308_0
tbb=2021.8.0=hdb19cb5_0
terminado=0.17.1=py311h06a4308_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=py311h06a4308_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchvision=0.15.2=py311_cpu
tornado=6.3.2=py311h5eee18b_0
tqdm=4.66.1=pyhd8ed1ab_0
traitlets=5.7.1=py311h06a4308_0
typing-extensions=4.7.1=py311h06a4308_0
typing_extensions=4.7.1=py311h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py311h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py311h06a4308_1
websocket-client=0.58.0=py311h06a4308_4
wheel=0.41.2=py311h06a4308_0
widgetsnbextension=4.0.5=py311h06a4308_0
xz=5.4.2=h5eee18b_0
y-py=0.5.9=py311h52d8a92_0
yaml=0.2.5=h7b6447c_0
ypy-websocket=0.8.2=py311h06a4308_0
zeromq=4.3.4=h2531618_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
